move state and logic of the full screen palyer to with-video hoc
